Move more modules around
