+++ jax/scipy/optimize/_line_search.py
@@ -1,7 +1,8 @@
+from typing import NamedTuple
+
 import jax.numpy as jnp
 from jax import value_and_grad
 from jax.lax import while_loop
-from typing import NamedTuple
 
 
 class LineSearchResults(NamedTuple):
