+++ jax/scipy/optimize/_minimize.py
